Move these checks to utils
